+++ src/Exception/Client/ClientException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace GraphQL\Exception\Client;
+
+/**
+ * Exception when a client error is encountered (4xx codes)
+ *
+ * Class ClientException
+ * @package GraphQL\Exception
+ */
+class ClientException extends BadResponseException
+{
+}