+++ src/Exception/Client/ServerException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace GraphQL\Exception\Client;
+
+/**
+ * Exception when a server error is encountered (5xx codes)
+ *
+ * Class ServerException
+ *
+ * @package GraphQL\Exception\Request
+ */
+class ServerException extends BadResponseException
+{
+}