+++ src/Exception/InvalidArgumentException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace GraphQL\Exception;
+
+use InvalidArgumentException as BaseInvalidArgumentException;
+
+/**
+ * Class InvalidArgumentException
+ *
+ * @package GraphQL\Exception
+ */
+class InvalidArgumentException extends BaseInvalidArgumentException implements Exception
+{
+}