+++ tests/QueryBuilderTest.php
@@ -7,8 +7,6 @@ use GraphQL\InlineFragment;
 use GraphQL\Query;
 use GraphQL\QueryBuilder\QueryBuilder;
 use GraphQL\RawObject;
-use GraphQL\Variable;
-use PHPUnit\Framework\TestCase;
 
 /**
  * This test case is responsible for testing the QueryBuilder and AbstractQueryBuilder classes